vuex -> typed-vuex
