+++ client/components/containers/card/ArtistCard.vue
@@ -51,7 +51,7 @@
 
 <script lang="ts">
 import Vue, { PropType } from 'vue';
-import { RootState } from 'vuex';
+import { RootState } from 'typed-vuex';
 
 import UserAvatar, { MediaIcon } from '~/components/parts/avatar/UserAvatar.vue';
 import { getImageSrc } from '~/scripts/converter/getImageSrc';
