+++ client/components/containers/card/PlaylistCard.vue
@@ -54,7 +54,7 @@
 
 <script lang="ts">
 import Vue, { PropType } from 'vue';
-import { RootState } from 'vuex';
+import { RootState } from 'typed-vuex';
 
 import ReleaseArtwork, { MediaIcon } from '~/components/parts/avatar/ReleaseArtwork.vue';
 import { getImageSrc } from '~/scripts/converter/getImageSrc';
