+++ client/components/containers/player/SeekBar.vue
@@ -29,7 +29,7 @@
 <script lang="ts">
 import Vue from 'vue';
 import debounce from 'lodash/debounce';
-import { RootState, ExtendedMutationPayload } from 'vuex';
+import { RootState, ExtendedMutationPayload } from 'typed-vuex';
 
 import { mssTime } from '~~/utils/mssTime';
