+++ client/pages/account/index.vue
@@ -33,7 +33,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import { RootState } from 'vuex';
+import { RootState } from 'typed-vuex';
 
 import TwoColumnsListCard, { ItemList } from '~/components/parts/list/TwoColumnsListCard.vue';
