+++ client/plugins/vuex.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-param-reassign */
 import { Plugin } from '@nuxt/types';
-import { ExtendedSubscribe } from 'vuex';
+import { ExtendedSubscribe } from 'typed-vuex';
 
 const injectVuex: Plugin = (context, inject) => {
   inject('state', () => context.store.state);
