+++ client/store/browse/getters.ts
@@ -1,4 +1,4 @@
-import { Getters } from 'vuex';
+import { Getters } from 'typed-vuex';
 import { BrowseState } from './state';
 
 export type BrowseGetters = {}
