+++ client/store/library/tracks/getters.ts
@@ -1,4 +1,4 @@
-import { Getters } from 'vuex';
+import { Getters } from 'typed-vuex';
 import { LibraryTracksState } from './state';
 
 export type LibraryTracksGetters = {
