+++ client/store/player/mutations.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign */
-import { Mutations } from 'vuex';
+import { Mutations } from 'typed-vuex';
 import { PlayerState } from './state';
 import type { SpotifyAPI, ZeroToHundred } from '~~/types';
