+++ client/store/search/mutations.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-param-reassign */
-import { Mutations } from 'vuex';
+import { Mutations } from 'typed-vuex';
 
 import { SearchState } from './state';
 import { SpotifyAPI } from '~~/types';
