+++ types/typed-vuex/index.d.ts
@@ -1,0 +1,2 @@
+import './vuex';
+import './impl';
